Safari-611.1.10/LayoutTests/fast/dom/NodeList/nodelist-name-getter-properties.html

<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<div id='aWildId'>Wild</div>

<script>
description('Test the JavaScript property descriptor of the name getter.');

var NodeList = document.getElementsByTagName('div');

shouldBeTrue("Object.getOwnPropertyDescriptor(NodeList, 'aWildId').configurable");
shouldBeFalse("Object.getOwnPropertyDescriptor(NodeList, 'aWildId').enumerable");
shouldBeFalse("Object.getOwnPropertyDescriptor(NodeList, 'aWildId').writable");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
